+++ shardis-ui/src/main/frontend/app/services/auth.service.ts
@@ -1,17 +1,19 @@
 import {Injectable} from '@angular/core';
 import {Http, Headers} from '@angular/http';
+import {LocalStorage} from 'ng2-webstorage/index';
 
 @Injectable()
 export class AuthService {
 
   private authenticated:boolean = false;
-  private tokenData:Oauth2TokenData = new Oauth2TokenData();
   private tokenExpirationDate:Date = null;
   private userData:any = null;
 
+  @LocalStorage()
+  private tokenData:Oauth2TokenData;
+
   constructor(public http:Http) {
-    if (window.localStorage.getItem('tokenData')) {
-      this.tokenData = JSON.parse(window.localStorage.getItem('tokenData'));
+    if (this.tokenData && this.tokenData.access_token) {
       this.authenticated = true;
       this.userData = this.decodeAccessToken(this.tokenData.access_token);
       this.tokenExpirationDate = new Date(this.userData.exp * 1000);
@@ -26,40 +28,49 @@ export class AuthService {
     return this.authenticated;
   }
 
-  public authenticate():void {
+  public authenticate(username: string, password: string) :Promise<string> {
+
     console.log('Authentication pending...');
 
-    var username = 'admin';
-    var password = 'xxxxxx';
+    return new Promise<string>((resolve, reject) => {
 
+      if (!username.trim()) {
+        reject('Username cannot be blank');
+      }
+      if (!password.trim()) {
+        reject('Password cannot be blank');
+      }
 
-    var basicAuthHeader = btoa(`acme:acmesecret`);
+      var basicAuthHeader = btoa(`acme:acmesecret`);
 
-    var headers = new Headers();
-    headers.append('Authorization', `Basic  ${basicAuthHeader}`);
-    headers.append('Accept', `application/json`);
-    headers.append('Content-Type', `application/x-www-form-urlencoded`);
+      var headers = new Headers();
+      headers.append('Authorization', `Basic  ${basicAuthHeader}`);
+      headers.append('Accept', `application/json`);
+      headers.append('Content-Type', `application/x-www-form-urlencoded`);
 
-    var data = 'username=' + encodeURIComponent(username) + '&password='
-      + encodeURIComponent(password) + '&grant_type=password';
+      var data = 'username=' + encodeURIComponent(username) + '&password='
+        + encodeURIComponent(password) + '&grant_type=password';
 
-    this.http.post('/auth/oauth/token', data, {headers: headers})
-      .subscribe(
-        data => {
-          this.tokenData = data.json();
-          window.localStorage.setItem('tokenData', JSON.stringify(this.tokenData));
-          this.authenticated = true;
-          this.userData = this.decodeAccessToken(this.tokenData.access_token);
-          this.tokenExpirationDate = new Date(this.userData.exp * 1000);
-        },
-        err => {
-          console.log(err);
-        }
-      );
+      this.http
+        .post('/auth/oauth/token', data, {headers: headers})
+        .subscribe(
+          data => {
+            this.tokenData = data.json();
+            this.authenticated = true;
+            this.userData = this.decodeAccessToken(this.tokenData.access_token);
+            this.tokenExpirationDate = new Date(this.userData.exp * 1000);
+            resolve('OK');
+          },
+          err => {
+            console.log(err);
+            reject('Username and password doesn\'t match');
+          }
+        );
+
+    });
   }
 
   public logout():any {
-    window.localStorage.removeItem('tokenData');
     this.tokenData = new Oauth2TokenData();
     this.userData = null;
     this.authenticated = false;