+++ shardis-ui/src/main/frontend/app/views/playground/heroes/app.component.ts
@@ -5,10 +5,12 @@ import {HeroService} from './hero.service.ts';
 
 @Component({
   selector: 'my-app',
-  template:`
+  template: `
+    <div style="text-align: center">
     <h1>{{title}}</h1>
     <h2>My Heroes</h2>
-    <ul class="heroes">
+    </div>
+    <ul class="heroes" style="width: 100%">
       <li *ngFor="let hero of heroes"
         [class.selected]="hero === selectedHero"
         (click)="onSelect(hero)">
@@ -16,17 +18,20 @@ import {HeroService} from './hero.service.ts';
       </li>
     </ul>
     <my-hero-detail [hero]="selectedHero"></my-hero-detail>
+
+
   `,
-  styles:[require('./app.component.scss')],
+  styles: [require('./app.component.scss')],
   directives: [HeroDetailComponent],
   providers: [HeroService]
 })
 export class AppComponent implements OnInit {
   title = 'Tour of Heroes';
-  heroes: Hero[];
-  selectedHero: Hero;
+  heroes:Hero[];
+  selectedHero:Hero;
 
-  constructor(private _heroService: HeroService) { }
+  constructor(private _heroService:HeroService) {
+  }
 
   getHeroes() {
     this._heroService.getHeroes().then(heroes => this.heroes = heroes);
@@ -36,6 +41,8 @@ export class AppComponent implements OnInit {
     this.getHeroes();
   }
 
-  onSelect(hero: Hero) { this.selectedHero = hero; }
+  onSelect(hero:Hero) {
+    this.selectedHero = hero;
+  }
 }