+++ shardis-ui/src/main/frontend/main.ts
@@ -23,7 +23,7 @@ require('./scss/main.scss');
  */
 export function main() {
 
-  return bootstrap(App, [
+  return bootstrap(App as any, [
     ...PROVIDERS,
     ...ENV_PROVIDERS,
     ...DIRECTIVES,
@@ -32,14 +32,6 @@ export function main() {
     .catch(err => console.error(err));
 
 }
-
-/*
- * Vendors
- * For vendors for example jQuery, Lodash, angular2-jwt just import them anywhere in your app
- * You can also import them in vendors to ensure that they are bundled in one file
- * Also see custom-typings.d.ts as you also need to do `typings install x` where `x` is your module
- */
-
 
 /*
  * Hot Module Reload