+++ shardis-ui/src/main/frontend/platform/browser/directives.ts
@@ -1,16 +1,14 @@
 import {provide, PLATFORM_DIRECTIVES} from '@angular/core';
+import {ROUTER_DIRECTIVES} from '@angular/router';
+import {MATERIAL_DIRECTIVES} from './material2/material2';
+import {RouterActive} from '../../app/directives/router-active.directive';
 
 // Angular 2 Router
-import {ROUTER_DIRECTIVES} from '@angular/router';
-import {FOUNDATION_COMPONENTS} from '../../app/components/foundation/accordion/foundation-components';
-
-// platform common directives
-import {RouterActive} from './router-active';
 
 // application_directives: directives that are global through out the application
 export const APPLICATION_DIRECTIVES = [
   ...ROUTER_DIRECTIVES,
-  ...FOUNDATION_COMPONENTS,
+  ...MATERIAL_DIRECTIVES,
   RouterActive
 ];