+++ shardis-ui/src/main/frontend/vendor.ts
@@ -5,7 +5,22 @@ import '@angular/core';
 import '@angular/common';
 import '@angular/http';
 import '@angular/router';
-
+// Material import 
+import '@angular2-material/button';
+import '@angular2-material/card';
+import '@angular2-material/checkbox';
+import '@angular2-material/core';
+import '@angular2-material/grid-list';
+import '@angular2-material/icon';
+import '@angular2-material/input';
+import '@angular2-material/list';
+import '@angular2-material/progress-bar';
+import '@angular2-material/progress-circle';
+import '@angular2-material/radio';
+import '@angular2-material/sidenav';
+import '@angular2-material/slide-toggle';
+import '@angular2-material/tabs';
+import '@angular2-material/toolbar';
 // RxJS
 import 'rxjs/Rx';